perf(home): update component request + abstracts pokemon-card
