+++ src/components/pokemon-list.tsx
@@ -1,112 +1,47 @@
-import { ReactNode, useEffect, useState } from 'react'
-import { Link } from 'react-router-dom'
+import { useQuery } from '@tanstack/react-query'
+import axios from 'axios'
+
+import { env } from '@/env'
 
 import pokeball from '../assets/pokeball.png'
 import { Loader } from './loader'
-import { Badge } from './ui/badge'
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from './ui/card'
-
-type PokemonListTypes = {
-  forEach(arg0: (pokemonUnit: any) => void): unknown
-  name: string
-  id: number
-  order: number
-  types: {
-    map(
-      arg0: (typeItem: {
-        type: { name: string }
-      }) => import('react/jsx-runtime').JSX.Element
-    ): ReactNode
-    name: string
-    url: string
-  }
-  sprites: {
-    other: {
-      home: {
-        front_default: string
-      }
-    }
-  }
-}
+import { PokemonCard } from './pokemon-card'
 
 export const PokemonList = () => {
-  const [pokemonList, setPokemonList] = useState<PokemonListTypes[]>([])
-  const [pokemonData, setPokemonData] = useState()
-  const [pokemonCount, setPokemonCount] = useState<number>(0)
+  /**
+   * 1) Create a function that will iterate between the url of each item in the array returned from the API.
+   * 2) Then, fetch those individual pokemon info and save it in the state.
+   */
+  const handlePokemonList = () => {
+    const endpoints = [] as string[]
 
-  const [isLoading, setIsLoading] = useState<boolean>(true)
-
-  useEffect(() => {
-    const getPokemonFromAPI = async () => {
-      const response = await fetch('https://pokeapi.co/api/v2/pokemon/')
-
-      const data = await response.json()
-
-      const { count, results } = data
-
-      setPokemonList(results)
-      setPokemonCount(count)
-    }
-
-    getPokemonFromAPI()
-  }, [])
-
-  useEffect(() => {
-    /**
-     * 1) Create a function that will iterate between the url of each item in the array returned from the API.
-     * 2) Then, fetch those individual pokemon info and save it in the state.
-     */
-    const listOfUrls = () => {
-      const pokemonUrls: string[] = []
-
-      if (pokemonList) {
-        pokemonList.forEach((pokemonUnit) => {
-          pokemonUrls.push(pokemonUnit.url)
-        })
-      }
-
-      return pokemonUrls
+    for (let index = 1; index < 152; index++) {
+      endpoints.push(`${env.VITE_API_BASE_URL}/pokemon/${index}/`)
     }
 
     const fetchPokemonData = async () => {
-      const listofPokemonUrls = listOfUrls()
+      const responses = await axios.all(
+        endpoints.map((endpoint) => axios.get(endpoint))
+      )
 
-      const promises = listofPokemonUrls.map(async (url) => {
-        const response = await fetch(url)
-
-        return response.json()
-      })
-
-      return Promise.all(promises)
+      const pokemonData = responses.map((response) => response.data)
+      return pokemonData
     }
 
-    if (pokemonList.length > 0) {
-      fetchPokemonData()
-        .then((data) => {
-          return setPokemonData(data)
-        })
+    return fetchPokemonData()
+  }
 
-        .catch((error) => console.error(error))
-
-        .finally(() => setIsLoading(false))
-    }
-  }, [pokemonList])
-
-  // console.log(pokemonList)
+  const { data: pokemonList, isLoading } = useQuery({
+    queryKey: ['pokemonList'],
+    queryFn: handlePokemonList,
+  })
 
   return (
-    <section className="mx-auto max-w-[1400px]">
+    <section className="mx-auto w-full max-w-[1400px]">
       <div className="flex w-full justify-end pb-4">
         <p className="flex items-center justify-end gap-1">
           <img src={pokeball} alt="" />
-          <span className="flex-1 text-2xl">{pokemonCount}</span>
+          <span className="flex-1 text-2xl">1302</span>
         </p>
       </div>
 
@@ -114,42 +49,8 @@ export const PokemonList = () => {
         <Loader />
       ) : (
         <div className="grid grid-cols-4 gap-4">
-          {pokemonData?.map((listItem) => {
-            return (
-              <Link key={listItem.id} to={`/pokemon/${listItem?.id}`}>
-                <Card className="border-white transition-[background] hover:bg-white/10">
-                  <img
-                    src={listItem?.sprites?.other?.home?.front_default}
-                    alt=""
-                  />
-
-                  <CardHeader className="mt-4 scroll-m-20 border-b border-t py-4 text-2xl font-semibold capitalize">
-                    <CardTitle className="capitalize tracking-tight">
-                      {listItem?.name}
-                    </CardTitle>
-
-                    <CardDescription className="tracking-tighter">
-                      #{listItem?.order?.toString().padStart(4, '0')}
-                    </CardDescription>
-                  </CardHeader>
-
-                  <CardFooter className="flex gap-2 py-4">
-                    {listItem?.types?.map(
-                      (typeItem: { type: { name: string } }) => {
-                        return (
-                          <Badge
-                            key={crypto.randomUUID()}
-                            className="capitalize text-white"
-                          >
-                            {typeItem.type.name}
-                          </Badge>
-                        )
-                      }
-                    )}
-                  </CardFooter>
-                </Card>
-              </Link>
-            )
+          {pokemonList?.map((listItem) => {
+            return <PokemonCard key={listItem.id} data={listItem} />
           })}
         </div>
       )}
